+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 
 ## Next Release
 
+#### Config
+
+- Add `changelog_paths` to config toml approach
+- Add `default_changelog_path` to config toml approach
+
 <!-- insertion marker -->
 ## [0.0.4](https://github.com/repo/owner/releases/tag/0.0.4) - 2025-12-20